requirements.txt
playwright>=1.44.0
python-docx>=1.1.0
openpyxl>=3.1.0
requests>=2.31.0
gspread>=6.0.0
google-api-python-client>=2.130.0
google-auth-oauthlib>=1.2.0
msal>=1.28.0
